chore: create components for table header and body
