+++ src/components/Table/index.tsx
@@ -3,26 +3,17 @@ import React from 'react';
 // Models
 import { Props, GenderList } from '../../models';
 
+// Components
+import TableHeader from '../TableHeader';
+import TableContent from '../TableContent';
+
 import './style.scss';
 
-// Icon
-import arrow from '../../assets/images/arrow.png';
-
 // Utils
-import {
-  getGenderImage,
-  characterFilter,
-  characterSorter,
-  genderFilter,
-  getCharacterCount,
-  getTotalHieght,
-  sortTableDataBy
-} from '../../utils/table';
+import { characterFilter, characterSorter, genderFilter } from '../../utils/table';
 
 const Table: React.FC<Props> = ({ movieCharacterInfo, filterTableData, filterText, sortConfig, sortTableData }) => {
-  const nameSortConfig = sortTableDataBy('name', sortConfig!);
-  const genderSortConfig = sortTableDataBy('gender', sortConfig!);
-  const heightSortConfig = sortTableDataBy('height', sortConfig!);
+
   const movieInfo = movieCharacterInfo?.movie!;
   const characters = movieCharacterInfo?.characters
     .filter(character => characterFilter(character, filterText!))
@@ -61,82 +52,10 @@ const Table: React.FC<Props> = ({ movieCharacterInfo, filterTableData, filterTex
 
       <table className='movies-list__table'>
         <thead className='movies-list__table-header'>
-          <tr className='movies-list__table-header-row'>
-            <th className='movies-list__table-heading'>
-              <button
-                className='movies-list__table-sorter'
-                type='button'
-                onClick={() => sortTableData!(nameSortConfig)}
-              >
-                Name
-              </button>
-              {sortConfig?.key === 'name' && (
-                <span
-                  className={`movies-list__table-sorter-icon ${
-                    sortConfig.direction === 'dsc' && sortConfig.key === 'name' ? 'rotate' : ''
-                  }`}
-                >
-                  <img src={arrow} alt='' />
-                </span>
-              )}
-            </th>
-            <th className='movies-list__table-heading'>
-              <button
-                className='movies-list__table-sorter'
-                type='button'
-                onClick={() => sortTableData!(genderSortConfig)}
-              >
-                Gender
-              </button>
-              {sortConfig?.key === 'gender' && (
-                <span
-                  className={`movies-list__table-sorter-icon ${
-                    sortConfig?.direction === 'dsc' && sortConfig?.key === 'gender' ? 'rotate' : ''
-                  }`}
-                >
-                  <img src={arrow} alt='' />
-                </span>
-              )}
-            </th>
-            <th className='movies-list__table-heading'>
-              <button
-                className='movies-list__table-sorter'
-                type='button'
-                onClick={() => sortTableData!(heightSortConfig)}
-              >
-                Height
-              </button>
-              {sortConfig?.key === 'height' && (
-                <span
-                  className={`movies-list__table-sorter-icon ${
-                    sortConfig?.direction === 'dsc' && sortConfig?.key === 'height' ? 'rotate' : ''
-                  }`}
-                >
-                  <img src={arrow} alt='' />
-                </span>
-              )}
-            </th>
-          </tr>
+          <TableHeader sortConfig={sortConfig} sortTableData={sortTableData} />
         </thead>
         <tbody>
-          {characters?.map((character, index) => (
-            <tr className='movies-list__table-row' key={index}>
-              <td className='movies-list__table-content'>{character.name}</td>
-              <td className='movies-list__table-content'>
-                <img title={character.gender} src={getGenderImage(character.gender)} alt={character.gender} />
-              </td>
-              <td className='movies-list__table-content'>{character.height}</td>
-            </tr>
-          ))}
-          <tr className='movies-list__table-row'>
-            <td className='movies-list__table-content'>
-              <strong>Total Characters: {getCharacterCount(characters!)}</strong>
-            </td>
-            <td></td>
-            <td className='movies-list__table-content'>
-              <strong>{getTotalHieght(characters!)}</strong>
-            </td>
-          </tr>
+          <TableContent characters={characters} />
         </tbody>
       </table>
     </div>
