+++ src/models/props.interface.ts
@@ -1,5 +1,6 @@
 import { Film } from './film.interface';
 import { MovieCharacter } from './moviecharacter.interface';
+import { People } from './people.interface';
 
 export interface Props {
   moviesList?: Film[];
@@ -12,4 +13,5 @@ export interface Props {
   sortConfig?: { key: string; direction: string } | null;
   sortTableData?: Function;
   setError?: Function;
+  characters?: People[];
 }
